pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wirestudio"
version = "0.1.0"
description = "Agent-driven IoT device design tool that produces ESPHome YAML."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "wirestudio contributors" }]
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
    "jinja2>=3.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "httpx>=0.27",
]

[project.scripts]
wirestudio-generate = "wirestudio.generate.__main__:main"
wirestudio-api = "wirestudio.api.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["wirestudio*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
